Use signed widths consistently in CFGAS_TxtBreak.
CFGAS_TxtBreak::AppendChar_Combination() may return a negative width
for combining characters, so ensure it isn't cast back-and-forth to
size_t.

Bug: chromium:1333307
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/94230
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ xfa/fde/cfde_texteditengine.cpp
@@ -896,7 +896,7 @@ wchar_t CFDE_TextEditEngine::GetChar(size_t idx) const {
              : content_[gap_position_ + gap_size_ + (idx - gap_position_)];
 }
 
-size_t CFDE_TextEditEngine::GetWidthOfChar(size_t idx) {
+int32_t CFDE_TextEditEngine::GetWidthOfChar(size_t idx) {
   // Recalculate the widths if necessary.
   Layout();
   return idx < char_widths_.size() ? char_widths_[idx] : 0;
